pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "myadegari",email = "myadegari37@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "uuid (>=1.30,<2.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "toml (>=0.10.2,<0.11.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "python-telegram-bot (>=22.0,<23.0)"
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
packages = [
    { include = "src", from = "." }
]

[tool.poetry.scripts]
cleanup = "scripts.cleanup:remove_pycache"
bale = "scripts.bale:replace_api_url"
start = "src.main2:main"
pre-start = "scripts.cleanup:remove_pycache"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
autoflake = "^2.3.1"

[tool.isort]
profile = "black"

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true